*update) 添加mongo spread_rate字段,fix ui error

+++ vnpy/app/spread_trading/backtesting.py
@@ -667,7 +667,7 @@ class BacktestingEngine:
         offset: Offset,
         price: float,
         volume: float,
-        payup: int,
+        payup: float,
         interval: int,
         lock: bool,
         spread_rate: float,
